requirements.txt
absl-py==0.9.0
alabaster==0.7.11
asn1crypto==0.24.0
astor==0.7.1
astroid==2.0.4
astropy==3.1.2
astunparse==1.6.3
Babel==2.6.0
backcall==0.1.0
bleach==1.5.0
cachetools==4.1.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==7.1.2
cloudpickle==0.5.5
colorama==0.3.9
comet-git-pure==0.19.14
comet-ml==3.0.2
configobj==5.0.6
cryptography==2.3.1
cvxopt==1.2.3
cvxpy==1.1.1
cycler==0.10.0
decorator==4.3.0
definitions==0.2.0
docutils==0.14
ecos==2.0.7.post1
entrypoints==0.2.3
everett==1.0.2
future==0.18.2
gast==0.2.2
google-auth==1.17.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
heartpy==1.2.6
hmmlearn==0.2.3
hrv==0.2.3
hrv-analysis==1.0.3
html5lib==0.9999999
hyperopt==0.1.1
idna==2.7
imagesize==1.0.0
importlib-metadata==1.6.1
intel-openmp==2019.0
ipykernel==4.8.2
ipython==6.5.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.12.1
Jinja2==2.10
joblib==0.14.0
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
keyring==13.2.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.5.1
Markdown==3.2.2
MarkupSafe==1.0
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.3
mkl==2019.0
mkl-fft==1.0.4
mkl-random==1.0.1
mlxtend==0.17.0
mpmath==1.1.0
nbconvert==5.3.1
nbformat==4.4.0
netifaces==0.10.7
networkx==2.2
nltk==3.5
nolds==0.5.1
nose==1.3.7
numpy==1.18.5
numpydoc==0.8.0
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opt-einsum==3.2.1
osqp==0.6.1
packaging==17.1
pandas==0.25.3
pandocfilters==1.4.2
parso==0.3.1
patsy==0.5.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
protobuf==3.12.2
psutil==5.4.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.4.0
pycparser==2.18
pyflakes==2.0.0
pyfnnd==0.5
Pygments==2.2.0
pylint==2.1.1
pymongo==3.7.1
pyOpenSSL==18.0.0
pyparsing==2.2.0
PySocks==1.6.8
python-dateutil==2.7.3
pytz==2018.5
pywin32==223
PyYAML==3.13
pyzmq==17.1.2
QtAwesome==0.4.4
qtconsole==4.4.1
QtPy==1.5.0
regex==2020.5.14
requests==2.24.0
requests-oauthlib==1.3.0
rope==0.11.0
rsa==4.6
scikit-learn==0.22
scipy==1.4.1
scs==2.1.2
seaborn==0.9.0
sets==0.3.2
simplegeneric==0.8.1
six==1.15.0
sklearn==0.0
snowballstemmer==1.2.1
spectrum==0.7.5
Sphinx==1.7.7
sphinxcontrib-websupport==1.1.0
spyder==3.3.1
spyder-kernels==0.2.6
statsmodels==0.9.0
sympy==1.3
tensorboard==1.15.0
tensorboard-plugin-wit==1.6.0.post3
tensorflow==1.5.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.5.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
testpath==0.3.1
tornado==5.1
tqdm==4.46.0
traitlets==4.3.2
typed-ast==1.1.0
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.14.1
wfdb==2.2.1
win-inet-pton==1.0.1
wincertstore==0.2
wrapt==1.12.1
wurlitzer==1.0.2
xlrd==1.1.0
zipp==3.1.0
